chore: move address-related helpers and tests from lib to p2p::addr
